moving focus finding stabilization
